+++ embrace-android-sdk/src/main/kotlin/io/embrace/android/embracesdk/internal/api/delegate/TrackedExperimentImpl.kt
@@ -1,0 +1,9 @@
+package io.embrace.android.embracesdk.internal.api.delegate
+
+import io.embrace.android.embracesdk.experiments.TrackedExperiment
+
+internal class TrackedExperimentImpl(
+    override val id: String,
+    override val variant: String?,
+    override val startedAt: Long?,
+) : TrackedExperiment
