+++ embrace-android-sdk/src/main/kotlin/io/embrace/android/embracesdk/internal/api/delegate/TrackedFeatureFlagImpl.kt
@@ -1,0 +1,8 @@
+package io.embrace.android.embracesdk.internal.api.delegate
+
+import io.embrace.android.embracesdk.experiments.TrackedFeatureFlag
+
+internal class TrackedFeatureFlagImpl(
+    override val id: String,
+    override val startedAt: Long?,
+) : TrackedFeatureFlag
